[TMR] Regenerated and tested tmr_tmr_timer_mode_pic32mx_usb_sk3
MH3-10809

+++ apps/tmr/tmr_timer_mode/firmware/src/config/pic32mx_usb_sk3/peripheral/tmr/plib_tmr2.h
@@ -76,11 +76,11 @@ void TMR2_Start(void);
 
 void TMR2_Stop(void);
 
-void TMR2_PeriodSet(uint16_t);
+void TMR2_PeriodSet(uint32_t);
 
-uint16_t TMR2_PeriodGet(void);
+uint32_t TMR2_PeriodGet(void);
 
-uint16_t TMR2_CounterGet(void);
+uint32_t TMR2_CounterGet(void);
 
 uint32_t TMR2_FrequencyGet(void);
